backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.5.2
sqlalchemy==2.0.23
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.1  # Added for PyTorch/BERT
transformers==4.35.2  # Added for HuggingFace BERT (assuming available; if not, install manually)
joblib==1.3.2  # For saving models